% Auto-generated: do not edit by hand
\name{pageMenu}

\alias{pageMenu}

\title{PageMenu component}

\description{

}

\usage{
pageMenu(id=NULL, dummy=NULL, dummy2=NULL, loading_state=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{dummy}{Character. }

\item{dummy2}{Character. }

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional): determines if the component is loading or not
  - prop_name (character; optional): holds which property is loading
  - component_name (character; optional): holds the name of the component that is loading. Object that holds the loading state object coming from dash-renderer}
}
