% Auto-generated: do not edit by hand
\name{sidebar}

\alias{sidebar}

\title{Sidebar component}

\description{

}

\usage{
sidebar(id=NULL, urls=NULL, depth=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{urls}{Unnamed list. URLs}

\item{depth}{Numeric. depth}
}
