% Auto-generated: do not edit by hand
\name{sidebar}

\alias{sidebar}

\title{Sidebar component}

\description{

}

\usage{
sidebar(children=NULL, id=NULL, urls=NULL, depth=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Custom content in the "no results found" panel}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{urls}{Unnamed list. URLs}

\item{depth}{Numeric. depth}
}
